requirements.txt
djangorestframework
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.2.0
celery==4.4.0
Django==3.0.2
importlib-metadata==1.5.0
kombu==4.6.7
MarkupSafe==1.1.1
psutil==5.6.7
PyMySQL==0.9.3
pytz==2019.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.0
vine==1.3.0
zipp==2.1.0
